requirements.txt
notion-client
python-dotenv
colorama
requests
duckduckgo-search
